Scope provider auth and throttles by tenant
